Don't store in the image build dependencies
python and base-build are only required to build Node.js packages with
native dependencies.  However, Python is not required to run the server,
so it's useless once npm install is done.

This commit moves the installation of temporal dependencies to the RUN
command associated with `npm install`, and once it's successful, it
uninstalls the dependencies.

The image may have been reduced by almost 40 to 45 percent, so it makes
a lot of sense.