refactor: extract database opening into module
This commit removes the logic for opening the database file and making
sure that the directory exists, into a separate file that can be
imported, removing a lot of file opening logic from Makibot.ts.

Additionally, mkdirp is removed as a direct dependency because the logic
has been refactored so that it uses the fs.mkdir method, which already
has a recursive option since Node 10.