refactor: use Makibot in command and hook params
This commit replaces the usage of Commando.CommandoClient in the
constructor() for commands and hooks, and instead it accepts a Makibot
instance, which is a subclass of CommandoClient.

The goal is to have configurations and common methods connected to the
client itself rather than in the commands, such as accessing the Config
object rather than using process.env all the time.