refactor: use fancier modlog messages
This commit will refactor the messages sent by the roster service, so
that they use a Rich Embed rather than a simple string. The code has
been refactored into a structure called ModlogEvent, so that it can be
extended by further actions that would require an event to be sent to
the modlog.