feat: automatic moderation actions (#94)
* feat: add !warn command

When a mod issues the !warn command, it requires the name of the user to
warn. The bot will send a message in the public modlog channel, to
notify the user that has been warned about something that should not be
done. It also applies the warn role automatically.

* refactor: extract warn applier into lib function

This commit will move the contents of the warn command into a new
function exported in lib/. Thus, it will be possible to use this
business logic in other parts that are not the warn command.

* feat: add warn service to react to warnable messages

When a mod reacts with ⚠️ to a message, should be warned without
mentioning the user in a !warn command. If additionally a mod sends the
🗑 emoji, should delete the message.

* refactor: use Server class to access roles and channels

* fix: mods should not be able to warn mods

Please, discuss that in private.

* feat: send warn events to the private modlog

This commit will send a copy of the warned messages, or of the warn
events, into the private modlog, so that mods can peek at the reason or
message that was warned and probably deleted.

* refactor: use local variables