feat: disallow !helper to users who have been warned
After applying this commit, users who receive a warn and have the warn
role cannot use the !helper command to change their helper presence
status. Because !warn will now revoke the helper role automatically, the
consequence is that, if an user is warned either through use of !warn or
through use of the :warn: emoji, those users won't be able to become
helpers until the warn is revoked.