enh(verify): force users to wait before approval
After investigation, the 10 minutes rule that Discord provides to
disallow users to type messages after joining a server does not work
when the user has confirmed their phone number. It has been noted that
spammers, flooders and other kinds of trolls will often use this trick
to bypass the 10 minutes rule.

After applying this commit, it won't be possible for users to bypass the
required cooldown period in order to join the server by actually having
the user wait until the joinedTimestamp passes a threshold.