Remove ping command
Do not deploy this commit until the new metrics can be ingested. It will
remove the !ping command expecting that the metrics can now be poked
using the HTTP server added in the previous commit, rather than having a
process sending the `!ping` command over the #server-status channel
every 5 minutes.