chore(ops): setup a healthcheck engine (#151)
* Set up an HTTP server and install the healthcheck endpoint

Using /healthcheck, a text string is served over an HTTP connection
containing the current ping if the server is connected. If the server is
not online, the request returns a status code of HTTP 503.

* Remove ping command

Do not deploy this commit until the new metrics can be ingested. It will
remove the !ping command expecting that the metrics can now be poked
using the HTTP server added in the previous commit, rather than having a
process sending the `!ping` command over the #server-status channel
every 5 minutes.