fix: sometimes GuildMember#joinedAt yields null
It may be that not always Discord.js provides the joinedAt timestamp,
which causes some members to have a value of null. Not checking this
causes the server to die, as reported in #159.

This commit will add checks before using the joinedAt field. Manual
investigation in our #captchas and #modlog channels shows that usually
these people have stayed longer than 5 minutes anyway, so the commit
makes the server bypass the verification process when joinedAt is not
available and automatically treat the user as if they had passed more
than 5 minutes in the server anyway.

This behaviour could be reverted in the future if #9 gets implemented at
some point and Clank can keep a local record that does not depend on the
response from Discord.js API.

Additionally, this commit will rewrite the message that a user receives
when triggers the captcha before the verification period expires, using
gender neutral language.