feat: introduce tags
A tag is a wrapper around a specific setting, allowing quickly to get
and set the value of a setting without having to type the key or the
guild of the setting every time.

As a refactor, the Server Settings object is now using tags. In the
future, it will be possible for members and other API objects to also
own tags, allowing custom properties to be attached to other elements.