feat: use webhooks for modlog events
This commit changes the way moderation events are pushed into the modlog
channel. Instead of having a hot pointer into the moderation channel
and sending messages there, messages will now be pushed using a webhook.

Webhooks still receive the moderation event as an embed, but now that we
use webhooks, we can override the avatar and nickname of the webhook.
This is used to use the moderation event title and icon as nickname and
avatar for the event.

Settings now don’t have to contain the name of the channel to use for
moderation events. Instead, servers will now have to contain the ID and
the token to use so that the webhook client can successfully send the
events.