Add antiraid toggle
This commit adds a toggle called AntiRaid to the Discord client. When
this toggle is enabled, the bot enters stealth mode, which is denoted by
the DND toggle being enabled (and persisted among reboots).

A getter is added to the client so that in the future commands and hooks
can get the status of the antiraid toggle. A command is also added, so
that mods can toggle on and off the antiraid mode.