feat: add antiraid mode (#177)
This commit adds a toggle called AntiRaid to the Discord client. When
this toggle is enabled, the bot enters stealth mode, which is denoted by
the DND toggle being enabled (and persisted among reboots).

When antiraid mode is enabled, the bot behaves diferently, such as
rejecting requests to validate an account if neccessary.

Note that at the moment the antiraid mode is global.