feat: add karma bag
The karma bag is a data structure that can hold points attached to the
given member. These points can be awarded in some cases such as sending
messages and receiving reactions, an can be taken in some cases such as
removing messages or receiving a warn or an alert.

The karma bag is designed to prevent farming and not more than 30 points
can be given per 24 hours to the same member. To achieve this safety, it
is important that the points are only given or taken from the karma bag
using the provided methods in the Karma class: grantPoints and
takePoints.