feat: add tombstones
Tombstones are special messages sent by the bot when a message is delete
from a channel that hadn't received new messages for at least a day.
These situations can confuse some users because an unread badge is still
retained by other users browsing the server, and they could believe that
the message is fresh while the message is actually old.

When a message is deleted from a channel, it is tested if the previous
message is older than 24 hours and in that case, a tombstone message is
sent to the channel. If the message is not that old, nothing is sent.

When a message is posted to the channel, the tombstone is deleted
automatically to avoid keeping in place blocking the conversation.
Tombstones are stored so that even if the client restarts, the bot can
still delete the message.