feat: add hook registry
Hooks are growing and there are more hooks. Instead of manually
instancing them in the Makibot class, this commit will add a registry,
similar to the Command Registry in use by Discord.js-Commando, so that
every .js or .ts file in the hooks/ directory is automatically loaded
when the bot starts.