feat: add Counter
Counter is a variation on Tags. They can only store numbers, but
additionally to methods to get, set and delete the value of a Counter,
there are helper methods designed to quickly increment and decrement the
value of a counter.

Counters may have a minimum and a maximum value, and every operation
that updates the value of a counter is able to clamp if neccessary the
actual persisted value according to the bounds in place.