feat: add TagBag
TagBag is a wrapper for Tags and Counter that automatically binds tags
and counters to a specific resolvable, such as a member or a server.

This allow client code to have an easier way to get, set and delete the
value of tags bound to GuildMembers, letting the TagBag expand the
SettingProvider keys with the guild ID and the guild member ID. Other
kinds of resolvables are also available, such as Servers, having a quick
way to get, delete and set the value of specific properties for Servers.

Accessors are added to the Server and Member classes for even faster
access to these properties.