feat: add expirable Tags
This commit will add a feature to tags that allow them to have a TTL
(time to live). If a tag has a TTL set, it is considered that the tag
can expire.

Tags that can expire will lose their value after the number of seconds
indicated by the TTL param pass since the last time they are updated.
Calling `#get` after this point will not retrieve the proper value.