feat: add roles when members join voice channels
This commit will make Clank add and remove roles to users that join
previously configured voice channels. This way, it is possible to have
an user receiving roles whenever they join a voice channel, and have it
removed when they leave a voice channel.

This allows, for instance, to make some channels only be visible as long
as the user is part of the voice channel. Members not in the voice
channel will not be able to see the channel because they don't have the
role.