feat: allow restarting a hook
This commit adds a watchdog to the HookManager so that hooks can be
registered if they support restarting. Whenever the makibot:restart
event is received by Makibot, if the paramter given is registered in the
HookManager as a hook that supports to be restarted, it will be
triggered so that the config can be reloaded.