feat: grant points for every message sent
In order to godfather old accounts with the correct amount of points, it
will be extremely difficult to scrap the search looking for the exact
amount of "awardable" points based on the user history.

If Clank had always stored a history of messages, this would have been
possible, but I never did this because I don't think this would be seen
as a correct move by the terms of use, the fact that I'm storing a copy
of all the messages received in the server.

So this commit will change the way the karma is given, by awarding one
point per message, no matter how many seconds since the last time points
were given.