feat(karma): add karma backlog
This commit adds a new tag to members called karma:offset which allows
giving members a backlog of karma, which is included when computing the
level. The amount of backlog can be controlled using the !setkarma
command.

It allows to offset the karma of godfathered accounts that were in the
system before the karma was added, so that we can grant these members
their real scores.