feat: handle karma using a slash command
This commit will replace the !karma command with a new slash command
called /karma, that returns the karma and level for the guild member
that issues the karma.

This is the first slash command so this commit will also add a simple
module with a handler for the INTERACTION_CREATE websocket event. This
is not Discord.js 13, so things are done manually. Once Discord.js 13 is
released maybe better integrations can be written.

Fixes https://github.com/makigas/clank/issues/261