feat: delete remaining commands
Since the pin and the modlog can already be controlled using the
settings endpoint in the REST API, there is no need to keep these
commands anymore.

Since those are the last commands existing yet, this commit also removes
the help command and disables registration of the commands/ directory,
because it is empty.