feat: add a function to build message embeds
The embed contains the contents of the message, as well as a link to go
back to the original message, and metadata such as the author of the
message and a timestamp at which the message was posted.

This creates a general purpose message quoter that can be used in the
pin system, or to custom embed messages, such as cross-posting messages.