refactor: use the new message quote in pins
This commit will change the behaviour of the pin hook to use the new
message quoter that was added in my previous commit. Unnecessary code
has been removed.

Additionally, this commit will fix an issue where reactions made in a
news group would not be sent to the pinboard channel, since the channel
of those messages is not of kind 'text' but 'news'.