feat: reply with toasts in hooks
Toasts are notifications built using embeds, that have a uniform look
and feel and consistent style. It also wraps the parameters so that in
the future, upgrading the bot to newer Discord.js versions organizes
better the breaking changes to avoid having to rewrite code in a lot of
areas.