fix: welcome messages are not being presented
Because karma is adjusted for everyone, new members get upgraded to
level 1 before the "welcome message" can detect whether the level was
zero.

This commit will amend that by forcing the karma upgrade to only be done
if the level is higher than zero, allowing new accounts to skip this
step.