feat: new response embeds for hooks and interactions (#394)
* feat: reply with toasts in hooks

Toasts are notifications built using embeds, that have a uniform look
and feel and consistent style. It also wraps the parameters so that in
the future, upgrading the bot to newer Discord.js versions organizes
better the breaking changes to avoid having to rewrite code in a lot of
areas.

* feat: wrap interaction responses in embeds too

This commit will make changes to the Interactions client so that the
commands generate embeds as response to actions instead of presenting
the message responses in plain text.

This doesn't affect pre-generated responses using the responder and
desresponder commands.