fix: fails to create settings on new bots
This commit will fix the errors that are happening when running the bot
in an environment where the settings file is either missing or the guild
is not present, due to the class failing to check for the presence of
the guild before accessing the cache.

Let's make use of the new ||= operator added in Node 16.