refactor: rewrite interactions using v9 API
This commit makes use of the new classes added to Discord.js 13 such as
the new events for the interactions and the CommandInteraction class,
and removes my custom glue code.

A loader is installed, so the bot will parse every command present in
the interactions/commands automatically. This will allow to create more
interactions in the future without having to register them manually.