fix: handle deleted messages and users in hooks
It could happen that triggering a tombstone for deleted messages causes
the application to crash. Similar, it might happen that a Member object
is created for an user that has left the server. Because the API will
return HTTP 404 in each of these cases, an error would be raised.

This commit adds some error handling to these hooks to prevent them
failing.