fix: prevent double warns
If the member to be warned is already warned, this commit will
change the behaviour of the applyWarn function so that no
double warns are triggered.

I found that when spambots are warned, they will keep sending
spam in channels designated for human appeals, so this will
cause the antispam system to continue being triggered.