feat(antiflood): ban muted members that trigger antiflood
Most of the time, if a muted member triggers again the antiflood is
because they are copy and pasting the message in channels reserved for
mute appeals. In this case, there isn't anything else to do. Either is a
bot, or a dumb.

Fixes https://github.com/makigas/clank/issues/443