fix(voicerole): do not trigger state changes on mute
This commit updates to discordjs-voice-role 2.0.2 to fix an issue where
users would still lose and regain their roles for a moment whenever they
mute or deafen. This causes noise and makes the nicknames flash in the
screen if the role is coloured.

Also, additional logic to early drop without logging to stdout is added.

Fixes: https://github.com/makigas/clank/issues/445
Pulls: https://github.com/makigas/discordjs-voicerole/releases/tag/v2.0.2