fix: leaving GuildMembers are supposed to be partial
This commit will not try to fetch a member that just left the
server, since you are not supposed to .fetch() them. Discord.js
still types the member as GuildMember | PartialGuildMember, but
since I cannot unpartial a guild member that leaves, I'm just
going to coerce leave events to PartialGuildMember.

As a consequence, I'm forced to remove the Joined at from the
event that is sent to the modlog, since it is typed as never in
PartialGuildMember for some reason, despite I'm able to see it
when I console.log() the payload.

Additionally, it seems that you should use GUILD_MEMBER as a
partial too if you want to be notified about left members in
some cases.

The remains of the PR are just cosmetic changes to please eslint
that were not handled before.