feat: add ctl for getting and putting KV settings
This commit adds three commands to clankctl:

* clankctl get-tag, to get the value of a tag.
* clankctl put-tag, to set the value of a tag.
* clankctl delete-tag, to delete the value of a tag.

They are designed to be used through the command line in case something
has to be debugged ever and the value of a specific key has to manually
change, overriding any kind of abstraction built on top of the setting
provider.

Using this endpoint it is possible to get the specific value of a tag,
no matter the preffix it is ussing. It is also possible to change
manually the value of a tag, and to delete it. It avoids having to rely
on manually touching the SQLite, given that the tags are stored as a big
JSON blob anyway.