feat: add button handler to see karma
When the button interaction for "karma_button" is received,
this handler will reply with the same information as the
Karma command.

Additionally, I changed the look and feel of the karma thing
to look nicer, and refactored the code so that it is not
duplicated three times in the system.