feat: add hook for logging deleted messages
This commit will log deleted events in the modlog, in order to
keep track of messages deleted during moderation, for auditing
purposes.

TODO: Who deleted the message? Different retention rules should
be applied. If the audit log says that the message was deleted
by a mod, it should be kept until the deletion is covered.
If it is deleted by the self-user, should be deleted as soon
as it is seen.