feat: design thread-only channels
This commit allows channels to be marked as thread-only, either
via a linkchannel or via a threadchannel. In both cases,
communication should be enforced using a thread.

Link channels will only accept messages that contain a link. If
a message that lacks a link is posted, the system will delete it.
This is useful for our links channel.

Thread channels will not make any test: any new message received
will spawn a thread for making comments. This is useful for our
server feedback channel and our showcase channel, to prevent
noise.