feat: add commands for threads management
This commit will add the following commands:

- /archivar can be used to archive a thread.
- /renombrar can be used to change the name of a thread.

In both cases, it is required that either the thread was created
by the author issuing the command, or that the thread was
started on top of a message created by the person issuing the
command.

This is useful when someone replies to your message, or when
the bot spawns a thread starting from a message you have
created.