refactor: reorganize and document HookManager
This commit will replace the old big fat constructor method for the
HookManager with one function per event listener. This makes the
cyclomatic complexity metric for the lib/hooks.ts function change from
"Bloody hell" to "Everything is good".

Also, I'm using this to document the behaviour of the hook manager in
the docs that were previously added to the project.