refactor: new mass-requirer for the hook system
The purpose of this refactor was to move the "requireAll" boilerplate
used to load all the files in a directory at the same time into a
separate module that can be refactored later to be used in interactions
as well.

Additionally, this lets me play with the new `unknown` type, introduced
in TypeScript 4.4, and refactor the type system so that there are less
type workarounds and make things more concise.