feat: detect airdrop spammers
This commit will add a new heuristic that will attempt to filter many
airdrop spammers that follow the same patterns (their messages always
mention everyone and have an URL, plus some keywords).

This heuristic is using a new Hook API callback: onMessagePremoderate.
As covered in the docs, any service implementing this callback will use
it to evaluate a message and automatically pick a moderation reason if
the message is not appropiate (triggers antiflood, triggers antispam).
Safe messages should return null.