refactor: migrate modlog to webhooks
This commit finalizes the work previously done for the ModlogV2
and migrates other private modlog events such as roster or delete
to webhooks.

It also takes the time to refactor the message templates to use
the newer features introduced recently in Discord, such as
native datetime templating (goodbye, humanize-duration), or using
the new builders provided in Discord.js to quickly build
mentions and links to channels and messages.

The new notifications are now part of the modlog directory, so
now I can start working on adding a file to lib/modlog/index to
build a facade, something that wasn't possible default because
there would be a conflict between lib/modlog/index and the old
lib/modlog.

Traces of the old modlog and webhook system are also gone, so
this commit will also remove the older key getters and setters
for the modlog webhook configuration from the RPC server. Old
getters are also removed from Makibot's Server class (Guild
wrapper).

Closes https://github.com/makigas/makibot/issues/426
Closes https://github.com/makigas/makibot/pull/447