chore(lint): pass eslint
After a lot of months of effort and making sure that the new code is
being introduced with lint-staged properly reporting OK, this commit
addresses the linting of the old areas of the code that still make
eslint fail, so that eslint starts passing.

Then, I configure a job in our CI workflow so that pull requests and
commits are always linted to make sure that no ugly code is added, ever
again.