feat: separate defaultMod from deleteMod
Deleted messages will go to a different moderation channel for deleted
messages, because I'm noticing that it is getting difficult to read the
modlog because of the amount of casual deletion messages received there.

For privacy reasons, the deletion channel should behave like a queue (or
a stack, which is what it is going to happen most probably). Messages
should be deleted once reviewed.