fix(antiflood): do not treat as flood the same channel
If a flooded message is sent to the same channel as last time, it should
not be considered flood. This is because the purpose of antiflood is to
prevent copying and pasting the same question to multiple channels, but
not to the same channel because that is not strictly speaking "copy to
multiple channels".

Also, there are some false positives for some people and maybe this will
help prevent the false positives to be triggered.

This commit will also modify the notification messages to make them more
straight, but also less scary.

Fixes https://github.com/makigas/makibot/issues/495