feat(linkchannel): try to get link title using Open Graph
When the linkchannel service opens a new thread for a link message, it
will first try to fetch the page contents and check whether there are
any valid Open Graph tags in the page that can be used to guess the
title of the page. This information will be presented as the default
thread name.