feat(antispam): disallow linking to RT.com
This commit will make the bot remove messages that contain links that
point to media that is directly funded by the Russian government.

These sites share propaganda that for some reason is sometimes
classified as "information" or "news", despite not being correct, as
some people may have noticed in the last week.

To keep things simple, the array uses strings and only parses domain
names, it matches both HTTP and HTTPS but it does not support wildcard
domains.

The feature is added to antispam to reuse the parsing, normalizing and
link detecting functions, although the function intentionally yields a
DELETE event and never returns anything else (so you cannot get banned
or muted by sharing these links).

On a less political plane, eventually this will be refactored to keep a
simple "blocklist" to be used for other spam airdroppers and
self-promotion platforms that doesn't depend on regexps to do its work.
(The always-DELETE outcome will not be refactored.)