feat(csgo): remove dead rules
Haven't seen a spambot using the old CS:GO trick in a Discord server in
a while, huh. I hope I don't have to regret this in the future, but this
commit will remove some sites from the main insta-mute list used to
catch and block general purpose spambots in Discord.

The block is lifted for prnt.sc, which as I have been told can also be
used for good. It is also lifted for general purpose sites ending with
.ru or .ru.com (as long as these domains do not match any other positive
rule that may cause the message to be moderated).

It also lifts the rules for the everyone + link check, since the airdrop
testing is already implemented as a separate function.