fix: autoarchive deleted linkchannel and threadchannel messages
When a message sent to a linkchannel or a threadchannel gets deleted,
the thread that is created after the original message should be archived
as well to hide it from the screen. The thread will be automatically
locked so that no one can send further messages. I considered deleting
the thread altogether, but I think that it is distasteful to those who
sent messages.

Fixes: https://github.com/makigas/makibot/issues/525