refactor: do not instantiate dependencies in Makibot constructor
Previously, makibot used the ready() hook to start the databases and all
the stuff, but this has the consequence of the bot being logged in
before being ready to use.

In this commit, the repositories used by Makibot will be provided as
parameters when spawning the Makibot class, which means that the
databases have to be initialised before calling Makibot. This has some
advantages:

* The bot will be ready by the time it logs in, so less errors.

* It is now easier to spawn the dependencies by themselves without
  having to spawn an entire bot.

* It is now easier to mock the dependencies.