feat: add trusted roles
Trusted roles is a dynamic setting that allows to configure some roles
that should be treated special, for instance, to prevent a ban or a warn
to be applied to some user that is marked as trusted.

This will also expose the list of roles in the API and will allow
makibotctl to manage the list of trusted roles.