feat: add snooze button on thread messages
When pressing this button, the bot will not send a tutorial on how to
use thread channels for 30 days since the button is pressed. Note that
you have to do this on your own button, you cannot just change the
settings for other people.

Fixes https://github.com/makigas/makibot/issues/553