refactor!: split guild and DM in command handling
The handle function of the command handler has been removed and instead
there are two optional functions now: handleDM and handleGuild. If you
want to implement a command that can be used in a guild, implement
handleGuild. If you want the command to be usable in DMs, implement
handleDM. You can implement both functions, one of them or none.

If a function is not implemented, triggering a command will layout an
error message claiming that the command cannot be used in that
circumstances. This error is handled outside the command.

BREAKING CHANGE: handle() is no longer invoked.