refactor!: split DM and group in button handler
The ButtonInteractionHandler class will now ignore the handle() function
and will instead try to use either the handleGuild() function or the
handleDM() function depending on the kind of location where the button
is.

Classes can implement none, one or both functions, and the interaction
handler will only delegate to the code if the function is implemented.
If a function is omitted, the interaction handler will render an error
claiming that the button cannot be pressed there.

BREAKING CHANGE: handle() is no longer invoked