refactor!: split context menu by type
The old ContextMenu handlers are gone and now there is one context menu
handler type depending on whether it is a message menu or a user menu.
It makes easier to fetch the messages or users assigned to a context
menu interaction.

It is also split by type of interaction: guild or DM, and answers
appropiately using the same new API as the other interaction types.