refactor!: rewrite pin as a context menu
The old pin hook (where reacting with a star to a message would send it
to the pinboard) has been removed and replaced with an interaction of
type Message Context Menu that, once triggered, will send a message to
the pinboard if the webhook is properly configured. The idea is that the
pinboard is now filled using webhooks rather than sending messages. It
is an interesting way to answer to the question: how should I refer to a
channel in the settings, by its name or by its ID? The answer: via its
webhook URL, of course!

Using the pinboard interaction will now copy the contents of the message
to the channel. At the same time, it will award 3 points to the user
that originally sent the message. Reacting with a star will have no
effect starting now.

BREAKING CHANGE: Removes the old pin system and all the functions.