feat: add server settings command
The /server command is restricted to server operators or moderators
depending on the context and allows to view and modify settings that are
otherwise meant to be accessed via makibotctl.

This command is not expected to replace makibotctl, but to make it
easier to view and update settings on the go.