feat(karma): add extra karma views
This commit will add a few extra view definitions that are required to
slice the data from the karma and offset table into something that will
evolve in the future.

* histogram: sums all the karma moved by day by account by type.
* transactions: a better computation model for bounties which includes
  fixing the bug of bounties not being capped at 300 per day.
* totals: the new way to compute the total amount of points of each
  user, using a single SQL statement rather than doing things
  client-side in Node.

Considering that for some reason node-sqlite only runs the last
statement in a batch, this commit splits the SQL commands into files,
and all the files are processed when the bot starts.