feat: replying to a message in a threadChannel should embed response
If someones ignore the way a thread channel works and replies directly
to an user in a thread based channel, the bot should embed the response.

Note that this will stay in operation a couple weeks since the
threadChannels will probably become deprecated in the upcoming weeks as
I manage how do the real forum channels will work.