refactor: deprecate antispam capabilities
Very thankful for the services provided by the bot in regards to this,
but Discord introduced this year a new automod that prevents these kinds
of messages from being sent even before they reach the server, so the
antispam measure provided by this bot is not needed anymore.

(Plus, is Discord doing things better? I haven't received an early
warning coming from the new automod in months)