chore: remove support for linkchannels and threadchannels
Discord has a forum feature since about a year or so, so there is no
need anymore to have channels designated as only threaded when we can
have native forums.

This commit removes the channel managers, any hook that monitors and
makes sure that rules are followed, and all the support features:
middlewares, the karma command not replying if it is a thread channel...

Also, /archivar and /renombrar are now removed as commands, because
native forums allow the OPs to rename and close their own threads
without requiring any extra commands.