chore(helper): remove helper support
This is an early feature that used to allow people to indicate that they
like helping other people when the help area of the server was used. The
support for mentioning Helpers role was removed a couple of years ago,
also their nicknames lost their yellow color a couple of years ago.

The native forum channels in Discord has also made this feature
obsolete.

Yada yada, removing this thing as well, and once deployed I can delete
the role.