refactor!: rewrite the SettingProvider
This commit finally removes the old SettingProvider that was inherited
from the discord.js-Commando and replaces it with a new one that uses
one row per setting.

The history of this class is very old. Initially the bot used the
SettingProvider that came for free with discord.js-Commando. However,
eventually discord.js-Commando got deprecated and stopped working with
recent Discord.js versions, so the bot had to move forward.

Because the bot used the SettingProvider for a lot of things, and there
were already hundreds of keys, a replica of the original SettingProvider
had to be built, with the exact same JavaScript API so that code could
continue to work, and the exact same database schema so that the
settings could still be read.

Since the bot eventually got tags and tagbags, fortunately the business
logic never uses directly the SettingProvider. However, it was still in
use.

One of the problems of the SettingProvider is that it is a mess. It is
actually a key-value database that is sharded by guild. However, the
original implementation just merges all the key-value into a big JSON
object and stores it into an SQLite database. This means that every time
you want to change or remove the value of a entry, you have to rebuild
a new JSON object and save it to database. This is resource-heavy.

The new implementation is still a SettingProvider, therefore it is
compatible with the code that uses this symbol, but instead every get,
set and delete uses a different SQL statement that actually uses a
separate row per entry. This means that instead of having a big JSON
object, now each tuple (guild, key) is its own row, with the value
stored on its own column. Changing the value of a setting means updating
a row, and deleting a setting means deleting a column.