refactor!: remove handleDM method
Commands and menus did not support DMs for a long time, but the code was
still there. Remove handleDM from the interaction manager and rename the
handleGuild method to just handle.

There is no need to check if it is a guild or a DM because the Discord
application settings in the developer panel now has a way to specify
whether the application can be used in DMs, guilds or both.