feat!: replace mod menu with "alert mods" feature
Moderators will not be able to use the mod menu anymore to handle a
message, and instead will require to use the native ban, kick or timeout
options available in Discord.

The reason for making this change is to start treating the moderation
features of Makibot as a passive consumer of moderation events, rather
than as an active moderation source. These days Discord has native
features to issue bans, timeouts and kicks, so Makibot should not try to
reinvent the wheel anymore, because there wheel is good enough.

The command has been renamed to "Alert mods" and it will always issue
the standard form to notify moderation about a message. In the future
the form will change into a modal so that the reporter can type whatever
they want.