refactor: make some files pass in strict mode
As a proof of concept, I want to make sure that npm run test is able to
build the code in strict mode. This already triggers some errors on some
files that are required during the test run. Not every file fails
because the coverage of this project is not high.

I've fixed those issues.

Also, I've fixed any issue in src/cmd/*.ts as reported by my text
editor.