refactor: better typing for src/lib/provider
After enabling strict mode, tsc complained about this package. It turns
out that the get() method for both the SettingProvider and the Tag were
not rich enough.

The old prototype was:

  get<T>(defVal?: T): Promise<T>

However, this fails because if defVal is not given, T is undefined. And
by changing the prototype to

  get<T>(defVal?: T): Promise<T | undefined>

Then every usage for this function complains that I am trying to use the
return without checking that it is not undefined, even when a
defaultValue is given.

The solution is to provide two overloads for the get method:

  get<T>(): Promise<T | undefined>;
  get<T>(defVal: T): Promise<T>;

This indicates that if there is a default value given for the get()
function, then it is impossible for the function to return undefined
(unless of course defVal is undefined). If a value is not given, then
sure, get() could return something or an undefined.

The method prototype is still

  get<T>(defVal?: T): Promise<T | undefined>

This is just part of the type checker.