feat!: passive modlog
The bot will not kick, ban or timeout anymore. These actions have to be
issued via the Discord web interface. The modlog will still listen to
these events and log them publicly and privately.

Since the modlog hook now can only detect timeouts that expire
naturally, I've renamed the hook to timeouts, in order to better capture
the intentions of the hook.

All the legacy code in src/lib/modlog has been deleted. Also, I've
refactored the code and all the units that are part of the modlog and
affected files so that they are conformant to strict TypeScript, a
requirement in order to upgrade the application.